# app/svc/docs/swagger.yaml
basePath: /
definitions:
...
  handler.AuthLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  handler.AuthRegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
...
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handler.OkResponse:
    properties:
      ok:
        type: boolean
    type: object
...
  handler.UsernameAvailableResponse:
    properties:
      available:
        type: boolean
    type: object
host: localhost:3000
info:
  contact: {}
  description: 'Backend-for-frontend: session management, campaigns, entries, annotations.'
  title: Quillit BFF API
  version: "1.0"
paths:
...
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Forwards credentials to auth-svc, creates an HTTP-only session
        cookie on success.
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.AuthLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Login
      tags:
      - auth
...
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Forwards credentials to auth-svc, creates an HTTP-only session
        cookie on success.
      parameters:
      - description: Registration credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.AuthRegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Register
      tags:
      - auth
...
  /api/auth/users/available:
    get:
      description: Proxies to auth-svc to check whether a username is free to register.
      parameters:
      - description: Username to check
        in: query
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UsernameAvailableResponse'
      summary: Username availability
      tags:
      - auth
...
swagger: "2.0"